Добавил JWT-токены и +зависимости от авторизации на категориях, проектах и get_user

+++ logic/category_routes.py
@@ -1,11 +1,13 @@
+from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
 from .categories import CategoryRepository, CategoryService
 from ..database import get_db
 from ..schemas.category import CategoryCreateSchema, CategoryUpdateSchema
+from .tokens import get_current_user
 
-router = APIRouter(prefix="/categories")
+router = APIRouter(prefix="/category", tags=["Categories"])
 
 # user_codes = {
 #     # "user_email": "code"
@@ -17,21 +19,25 @@ def get_category_repository(db: Session = Depends(get_db)):
 def get_category_service(user_repository: CategoryRepository = Depends(get_category_repository)):
     return CategoryService(user_repository)
 
+ServiceDependency = Annotated[CategoryService, Depends(get_category_service)]
+UserDependency = Annotated[dict, Depends(get_current_user)]
 
-@router.post("/add",
-             tags=["Categories"])
+@router.post("/")
 async def add_category(category_data: CategoryCreateSchema, 
-                       service: CategoryService = Depends(get_category_service)):
-    category = service.add_category(category_data)
+                       service: ServiceDependency,
+                       user: UserDependency):
+    user_id = user["user_id"]
+    
+    category = service.add_category(user_id, category_data)
 
     return {"id": category.id, "name": category.name, "color": category.color}
 
 
-@router.put("/update", 
-            tags=["Categories"])
+@router.put("/")
 async def update_category(category_id: int, 
                           category_data: CategoryUpdateSchema, 
-                          service: CategoryService = Depends(get_category_service)):
+                          service: ServiceDependency,
+                          user: UserDependency):
     category = service.update_category(category_id, category_data)
 
     if category is None:
@@ -39,11 +45,23 @@ async def update_category(category_id: int,
 
     return {"id": category.id, "name": category.name, "color": category.color}
 
+@router.get("/all")
+async def get_all_user_categories(service: ServiceDependency,
+                                  user: UserDependency):
+    user_id = user["user_id"]
+    categories = service.get_categories_by_user_id(user_id)
 
-@router.delete("/delete",
+    return {"categories": categories}
+
+
+@router.delete("/",
              tags=["Categories"])
 async def delete_category(category_id: int, 
-                          service: CategoryService = Depends(get_category_service)):
+                          service: ServiceDependency,
+                          user: UserDependency):
+    if not service.check_category_id(user["user_id"], category_id):
+        raise HTTPException(status_code=403, detail="Неправомерное удаление данных")
+    
     category = service.remove_category(category_id)
 
     if category is None:
